requirements.txt
discord.py
requests
py-cpuinfo
